Add native lazyhint, change script to setup linux

+++ setup.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env fish
+#!/usr/bin/env bash
 #
 # fractional scaling gnome :  gsettings set org.gnome.mutter experimental-features "['scale-monitor-framebuffer']" 
 
@@ -10,7 +10,17 @@ fi
 
 echo $XDG_CONFIG_HOME
 
-# sudo apt install ripgrep fd-find fzf zoxide python3-pip
+sudo apt install ripgrep fd-find fzf zoxide python3-pip fish git keychain
+
+if ! command -v starship &> /dev/null
+then
+    echo "starship could not be found"
+	curl -sS https://starship.rs/install.sh | sh
+fi
+
+if [ ! -d "${HOME}/.config" ]; then
+	mkdir "${HOME}/.config/fish"
+fi
 
 clone_path="${clone_path:-"${PWD}"}"
 echo "Current working directory: $clone_path"
@@ -20,16 +30,17 @@ ln -fs "${clone_path}/nvim" "${HOME}/.config/nvim"
 
 
 ln -fs "${clone_path}/starship/starship.toml" "${HOME}/.config/starship.toml"
+ln -fs "${clone_path}/fish/config.fish" "${HOME}/.config/fish/config.fish"
 
 # git
 ln -fs "${clone_path}/git/.gitconfig" "${HOME}/.gitconfig"
 ln -fs "${clone_path}/git/.gitignore" "${HOME}/.gitignore"
 ln -fs "${clone_path}/git/.gitattributes" "${HOME}/.gitattributes"
 
-cp "${clone_path}/git/.gitconfig.local" "${HOME}/.gitconfig.local" 
+if [ ! -f "${HOME}/.gitconfig.local" ]; then
+	echo "copying .gitconfig.local"
+	cp "${clone_path}/git/.gitconfig.local" "${HOME}/.gitconfig.local" 
+fi
 
-# New-Item -ItemType SymbolicLink -Force -Path $HOME/.gitconfig -Target $path\.gitconfig;
-# New-Item -ItemType SymbolicLink -Force -Path $HOME/.gitignore -Target $path\.gitignore
-# New-Item -ItemType SymbolicLink -Force -Path $HOME/.gitattributes -Target $path\.gitattributes
 
-
+./fish/setup.sh
